UX-565 Move extractChildren to helpers package; move tokens.js
